Various protection changes

+++ restaurant/OrderItem.java
@@ -1,14 +1,14 @@
 public class OrderItem {
 
     public String name;
-    public double price;
+    public float price;
 
     /**
      * Constructor method
      * @param name - the name of the ordered item
      * @param price - the price of the ordered item
      */
-    public OrderItem(String name, double price){
+    public OrderItem(String name, float price){
 
         this.name = name;
         this.price = price;
